refactor: 초기 데이터 파일 domain디렉터리 하위로 이동
초기 데이터를 상수라고 볼 수 없음.
domain중 일부라고 생각하여 이동하였음.
